simplificação dos direcionamentos

+++ demo/src/main/java/com/esig/br/demo/controller/TarefaController.java
@@ -1,6 +1,7 @@
 package com.esig.br.demo.controller;
 
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.annotation.PostConstruct;
@@ -8,7 +9,6 @@ import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 import javax.faces.view.ViewScoped;
 
-import org.ocpsoft.rewrite.annotation.Join;
 import org.primefaces.PrimeFaces;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -23,7 +23,6 @@ import lombok.Setter;
 
 @Component
 @ViewScoped
-@Join(path = "/nova-tarefa", to = "/tarefa-form.jsf")
 public class TarefaController implements Serializable {
     private static final long serialVersionUID = 1L;
     private static final String UPDATE_JSF_MENSSAGE = "form:messages";
@@ -37,12 +36,15 @@ public class TarefaController implements Serializable {
 
     @Getter @Setter
     private transient Tarefa tarefa = new Tarefa();
-    @Getter
-    private transient List<Tarefa> tarefas;
+    @Getter @Setter
+    private transient List<Tarefa> tarefas = new ArrayList<>();
+    @Getter @Setter
+    private transient List<Tarefa> tarefasFiltradas = new ArrayList<>();
     
     @PostConstruct
     public void loadData() {
         tarefas = tarefaRepository.findAll();
+        tarefasFiltradas.addAll(tarefas);
     }
 
     public String salvar() {
@@ -60,14 +62,18 @@ public class TarefaController implements Serializable {
     }
 
     public void deletar() {
-        tarefaRepository.delete(tarefa);
-        tarefas.remove(tarefa);
-        tarefa = new Tarefa();
-        FacesContext.getCurrentInstance().addMessage(null, new FacesMessage("Tarefa Excluida"));
-        PrimeFaces.current().ajax().update(UPDATE_JSF_MENSSAGE, UPDATE_JSF_LISTTABLE);
+        if(tarefa.getId()!=null){
+            tarefaRepository.delete(tarefa);
+            tarefas.remove(tarefa);
+            tarefasFiltradas.remove(tarefa);
+            tarefa = new Tarefa();
+            FacesContext.getCurrentInstance().addMessage(null, new FacesMessage("Tarefa Excluida"));
+            PrimeFaces.current().ajax().update(UPDATE_JSF_MENSSAGE, UPDATE_JSF_LISTTABLE);
+        }
     }
 
     public void concluir() {
+        if(tarefa.getSituacao()==null || tarefa.getSituacao().equals(Situacao.FINALIZADO)) return;
         tarefa.setSituacao(Situacao.FINALIZADO);
         tarefaRepository.save(tarefa);
         tarefa = new Tarefa();
